Rocky Dirt mining behavior
The block that you get from mining Rocky Dirt now depends on the tool being used. Breaking the block with a Pickaxe will make it drop Cobblestone and breaking it with a Shovel will give you Dirt. To get the block itself you need to use a tool with Silk Touch.